Unify GetIssueTypes and GetComponents
