+++ Atlassian.Jira/IIssueTypeService.cs
@@ -15,11 +15,5 @@ namespace Atlassian.Jira
         /// Returns all the issue types within JIRA.
         /// </summary>
         Task<IEnumerable<IssueType>> GetIssueTypesAsync(CancellationToken token = default(CancellationToken));
-
-        /// <summary>
-        /// Returns the issue types within JIRA for the project specified.
-        /// </summary>
-        Task<IEnumerable<IssueType>> GetIssueTypesForProjectAsync(string projectKey, CancellationToken token = default(CancellationToken));
-
     }
 }
