+++ Atlassian.Jira/Project.cs
@@ -105,9 +105,9 @@ namespace Atlassian.Jira
         /// Gets the issue types for the current project.
         /// </summary>
         /// <param name="token">Cancellation token for this operation.</param>
-        public Task<IEnumerable<IssueType>> GetIssueTypesAsync(CancellationToken token = default(CancellationToken))
+        public async Task<IEnumerable<IssueType>> GetIssueTypesAsync(CancellationToken token = default(CancellationToken))
         {
-            return _jira.IssueTypes.GetIssueTypesForProjectAsync(Key, token);
+            return (await _jira.Projects.GetProjectIssueTypesAndComponentsAsync(Key, token)).Item1;
         }
 
         /// <summary>
