+++ Atlassian.Jira/Remote/IssueTypeService.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Newtonsoft.Json;
 using RestSharp;
 
 namespace Atlassian.Jira.Remote
@@ -21,19 +19,6 @@ namespace Atlassian.Jira.Remote
         {
             var remoteIssueTypes = await _jira.RestClient.ExecuteRequestAsync<RemoteIssueType[]>(Method.GET, "rest/api/2/issuetype", null, token).ConfigureAwait(false);
             var issueTypes = remoteIssueTypes.Select(t => new IssueType(t));
-            return issueTypes;
-        }
-
-        public async Task<IEnumerable<IssueType>> GetIssueTypesForProjectAsync(string projectKey, CancellationToken token = default(CancellationToken))
-        {
-            var resource = String.Format("rest/api/2/project/{0}", projectKey);
-            var projectJson = await _jira.RestClient.ExecuteRequestAsync(Method.GET, resource, null, token).ConfigureAwait(false);
-            var serializerSettings = _jira.RestClient.Settings.JsonSerializerSettings;
-
-            var issueTypes = projectJson["issueTypes"]
-                .Select(issueTypeJson => JsonConvert.DeserializeObject<RemoteIssueType>(issueTypeJson.ToString(), serializerSettings))
-                .Select(remoteIssueType => new IssueType(remoteIssueType));
-
             return issueTypes;
         }
     }
