+++ Atlassian.Jira/Remote/ProjectService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
 using RestSharp;
 
 namespace Atlassian.Jira.Remote
@@ -27,6 +28,24 @@ namespace Atlassian.Jira.Remote
             var resource = String.Format("rest/api/2/project/{0}?expand=lead,url", projectKey);
             var remoteProject = await _jira.RestClient.ExecuteRequestAsync<RemoteProject>(Method.GET, resource, null, token).ConfigureAwait(false);
             return new Project(_jira, remoteProject);
+        }
+
+        public async Task<(IEnumerable<IssueType>, IEnumerable<ProjectComponent>)>
+            GetProjectIssueTypesAndComponentsAsync(string projectKey, CancellationToken token = default)
+        {
+            var resource = String.Format("rest/api/2/project/{0}", projectKey);
+            var projectJson = await _jira.RestClient.ExecuteRequestAsync(Method.GET, resource, null, token).ConfigureAwait(false);
+            var serializerSettings = _jira.RestClient.Settings.JsonSerializerSettings;
+
+            var issueTypes = projectJson["issueTypes"]
+                .Select(issueTypeJson => JsonConvert.DeserializeObject<RemoteIssueType>(issueTypeJson.ToString(), serializerSettings))
+                .Select(remoteIssueType => new IssueType(remoteIssueType));
+
+            var components = projectJson["components"]
+                .Select(componentJson => JsonConvert.DeserializeObject<RemoteComponent>(componentJson.ToString(), serializerSettings))
+                .Select(remoteComponent => new ProjectComponent(remoteComponent));
+
+            return (issueTypes, components);
         }
     }
 }
